+++ src/main/java/com/influy/domain/seller/service/SellerService.java
@@ -1,43 +1,17 @@
 package com.influy.domain.seller.service;
 
-import com.influy.domain.profileLink.entity.ProfileLink;
-import com.influy.domain.seller.converter.SellerConverter;
 import com.influy.domain.seller.dto.SellerRequestDTO;
 import com.influy.domain.seller.entity.ItemSortType;
 import com.influy.domain.seller.entity.Seller;
-import com.influy.domain.seller.repository.SellerRepository;
-import com.influy.global.apiPayload.code.status.ErrorStatus;
-import com.influy.global.apiPayload.exception.GeneralException;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
-@Service
-@Transactional(readOnly = true)
-@RequiredArgsConstructor
-public class SellerService {
+public interface SellerService {
+    Seller join(SellerRequestDTO.Join requestBody);
 
-    private final SellerRepository sellerRepository;
+    Seller getSeller(Long sellerId);
 
-    @Transactional
-    public Seller join(SellerRequestDTO.Join requestBody) {
-        Seller seller = SellerConverter.toSeller(requestBody);
-        return sellerRepository.save(seller);
-    }
+    Seller updateSeller(Long sellerId, SellerRequestDTO.UpdateProfile requestBody);
 
-    public Seller getSeller(Long sellerId){
-        return sellerRepository.findById(sellerId).orElseThrow(()->new GeneralException(ErrorStatus.SELLER_NOT_FOUND));
-    }
+    Seller updateItemSortType(Long sellerId, ItemSortType sortBy);
 
-    @Transactional
-    public Seller updateSeller(Long sellerId, SellerRequestDTO.UpdateProfile requestBody) {
-        Seller seller = getSeller(sellerId);
-        return seller.setProfile(requestBody);
-    }
-
-    @Transactional
-    public Seller updateItemSortType(Long sellerId, ItemSortType sortBy) {
-        Seller seller = getSeller(sellerId);
-        return seller.setItemSortType(sortBy);
-    }
+    void checkItemMatchSeller(Long sellerId, Long itemId);
 }
